+++ components/CcustomDropdown/CcustomDropdown.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables';
-@import '../../styles/mixins';
+@import 'styles/variables';
+@import 'styles/mixins';
 
 .custom-dropdown {
   display: flex;
