+++ components/Sspinner/Sspinner.scss
