+++ components/WweatherWidget/DdayCard/DdayCard.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables';
-@import '../../../styles/mixins';
+@import 'styles/variables';
+@import 'styles/mixins';
 
 .day-card-wrapper {
   background: $widget-bg;
