+++ components/WweatherWidget/WweatherWidget.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables';
-@import '../../styles/mixins';
+@import 'styles/variables';
+@import 'styles/mixins';
 
 .widget-loader {
   margin-top: 60px;
