+++ components/WweatherWidget/WwidgetBody/WwidgetBody.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables';
-@import '../../../styles/mixins';
+@import 'styles/variables';
+@import 'styles/mixins';
 
 .widget-body {
   &__spinner-wrapper {
