+++ next.config.js
@@ -1,5 +1,6 @@
 const withSass = require('@zeit/next-sass');
 const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
+const path = require('path');
 
 module.exports = withSass({
   webpack: config => {
@@ -8,6 +9,12 @@ module.exports = withSass({
       fs: 'empty',
     };
     config.plugins.push(new MomentLocalesPlugin());
+    config.resolve = {
+      alias: {
+        utils: path.resolve(__dirname, 'utils'),
+        styles: path.resolve(__dirname, 'styles'),
+      },
+    };
     return config;
   },
 });
